requirements.txt
django==4.1.3
celery==5.2.7
pre-commit==2.20.0
djangorestframework==3.14.0
